refactor(src): organize into guard/, job_lock/, utility/, scripting/ folders
Per the reorganization plan committed earlier
(.docs/plans/20260429-0810-src-reorganization.md). Pure mechanical
move + import-path rewrite + a clean split of utility.ts into two
semantic homes — no logic changes, no test changes (still 304/304
green).

New layout:

  src/
    api_server.ts              (bootstrap, unchanged)
    custom.d.ts
    logger.ts                  (cross-cutting, intentionally root)
    system_status.ts (+test)   (top-level state, intentionally root)

    guard/
      api_key_validator.ts (+test)
      write_guard.ts (+test)
      ws_lock_guard.ts (+test)   ← renamed from job_lock_middleware.ts
                                  (file now contains only WS-side
                                  rejectIfLocked, not Express middleware)

    job_lock/
      job_lock.ts (+test)
      job_lock.integration.test.ts

    utility/
      semver.ts (+test)
      short_id.ts (+test)        ← extracted from utility.ts;
                                  contains generateShortId only

    scripting/
      script_converter.ts (+test)
      script_duration.ts (+test) ← extracted from utility.ts;
                                  contains updateScriptDuration +
                                  calculateLengthDS

utility.ts is removed: its two semantic concerns split cleanly
(generateShortId is generic; updateScriptDuration + calculateLengthDS
operate on Script and belong with the other script logic). The
former utility.test.ts splits along the same lines.

Imports updated:
  - 3 stale-path imports in api_server.ts (guard/, job_lock/, etc.)
  - 1 stale-path import in dal/repositories/playlist_repository.ts
  - 1 stale-path import (split into 2) in
    dal/repositories/script_repository.ts
  - All moved files' internal `./X.js` references rewritten to
    `../X.js` for cross-folder paths
  - `git mv` preserves history for every renamed file

Verified: build clean (lint + tsc), 304/304 tests green, prettier
+ lint clean, no remaining stale imports per
git grep -nE "from 'src/(utility|semver|...)\.js'" — empty.

Plan: .docs/plans/20260429-0810-src-reorganization.md
