+++ astros_api/src/scripting/script_duration.ts
@@ -1,14 +1,4 @@
-import { Script } from './models/index.js';
-
-export function generateShortId(prefix: string): string {
-  const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz';
-  let result = `${prefix}${Math.floor(Date.now() / 1000000)}`;
-  const charactersLength = characters.length;
-  for (let i = 0; i < 3; i++) {
-    result += characters.charAt(Math.floor(Math.random() * charactersLength));
-  }
-  return result;
-}
+import { Script } from '../models/index.js';
 
 /**
  * Updates the duration of a script if it is undefined, null, NaN, or less than 0.
